+++ randovania/interface_common/persisted_options.py
@@ -148,6 +148,9 @@ def _convert_v24(options: dict) -> dict:
     # added Dread's missile pack recolor
     return options
 
+def _convert_v25(options: dict) -> dict:
+    # added Dread's auto tracker
+    return options
 
 _CONVERTER_FOR_VERSION = [
     None,
@@ -174,6 +177,7 @@ _CONVERTER_FOR_VERSION = [
     _convert_v22,
     _convert_v23,
     _convert_v24,
+    _convert_v25,
 ]
 _CURRENT_OPTIONS_FILE_VERSION = migration_lib.get_version(_CONVERTER_FOR_VERSION)